+++ src/main/java/com/inet/INETApplication.java
@@ -2,8 +2,10 @@ package com.inet;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.retry.annotation.EnableRetry;
 
 @SpringBootApplication
+@EnableRetry
 public class INETApplication {
 
 	public static void main(String[] args) {