+++ src/main/java/com/inet/service/SchoolService.java
@@ -4,13 +4,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import com.inet.entity.School;
 import com.inet.repository.SchoolRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
 @Service
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Log4j2
 public class SchoolService {
@@ -44,5 +46,9 @@ public class SchoolService {
     public void deleteSchool(Long id) {
         log.info("Deleting school with id: {}", id);
         schoolRepository.deleteById(id);
+    }
+
+    public Optional<School> findById(Long schoolId) {
+        return schoolRepository.findById(schoolId);
     }
 }